[HW09] Factor code processing slice
